requirements.txt
django==1.11.11
pymysql==0.9.2
celery==4.3.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
redis==3.3.10
python-nmap==0.6.1
pyvmomi==6.7.1
channels==2.3.0
channels-redis==2.3.3
paramiko==2.7.1

eventlet==0.25.1
uwsgi==2.0.18



